requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
argparse
nltk
scikit-learn
PyMuPDF
python-docx 